Fixed a few of bugs
When username and password are blank, the user is not allowed to
login/register/change password.
Edit the method ‘Add’ on the main window, in default situation, new
student will be added at the beginning of the list, if a row is
selected while adding a student, the new student will be added after
that line.